refine policy
